test(statedb): gate TestWatcherEventDedup behind hostsensitive build tag (#969)
This test races two goroutines on a shared SQLite handle. Under -race on
some hosts, kernel scheduling + busy-handler timing trips SQLITE_BUSY
non-deterministically. Move it to a hostsensitive-tagged file so the
default pre-push and CI runs stay green; opt in via:

    go test -tags hostsensitive -race ./internal/statedb/...

Refs: #969

Committed by Ashesh Goplani
